+++ CHANGELOG.md
@@ -13,6 +13,13 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- IndividualIdentity resource
+- IndividualDocument resource
+- IssuingEmbossingKit resource
+- CreditHolmes resource
+### Changed
+- cardDesingID and envelopID to kitID attribute
 
 ## [0.2.0] - 2022-12-12
 ### Added